+++ src/Protocol/Palettes/Palette_1_15.cpp
@@ -9653,7 +9653,7 @@ namespace Palette_1_15
 			case Statistic::WalkOnWaterOneCm: return 8;
 			case Statistic::WalkOneCm: return 5;
 			case Statistic::WalkUnderWaterOneCm: return 12;
-			default: return -1;
+			default: return UInt32(-1);
 		}
 	}
